+++ server/app.ts
@@ -4,30 +4,16 @@ import * as bodyParser from "body-parser";
 import Axios from "axios";
 
 // import some controllers; extract these to the controllers folder later
-import {
-  verifySession,
-  checkIfFriend,
-  checkIfFriendWithUserID,
-  getUserID,
-} from "./controllers/VerifySession";
+import { verifySession } from "./controllers/VerifySession";
 import { SessionType } from "./types/session";
 
 // import routes files
 import { contentRouter } from "./routes/ManageContent";
 import { viewContentRouter } from "./routes/ViewContent";
+import { manageUserRouter } from "./routes/UserManagement";
 
 import passport from "passport";
 import session from "express-session";
-import { FriendPairType, FriendRequestUserType } from "./types/content";
-import {
-  acceptRequest,
-  checkUsername,
-  createUser,
-  denyRequest,
-  getIncomingFriendRequests,
-  getOutgoingFriendRequests,
-  sendFriendRequest,
-} from "./controllers/UserManagement";
 
 // import packages for MySQl session store
 const mysql = require("mysql2/promise");
@@ -73,8 +59,7 @@ app.use(passport.session());
 app.use(bodyParser.json());
 app.use("/content", contentRouter);
 app.use("/view", viewContentRouter);
-
-// auth testing
+app.use("/user", manageUserRouter);
 
 const isLoggedIn = (req: Request, res: Response, next: NextFunction) => {
   req.user ? next() : res.sendStatus(401);
@@ -127,8 +112,8 @@ app.get("/verify", (req: Request, res: Response) => {
   });
 });
 
-app.post("/get-recipes", (req: Request, res: Response) => {
-  const recipeSearch = req.body.query;
+app.get("/recipes", (req: Request, res: Response) => {
+  const recipeSearch = req.params.query;
   let recipeResponse;
   const recipeOptions = {
     params: {
@@ -155,228 +140,12 @@ app.post("/get-recipes", (req: Request, res: Response) => {
     });
 });
 
-app.post("/check-if-friend-with-google-id", (req: Request, res: Response) => {
-  const reqUserID = String(req.user?.id);
-  const secondID = req.body.secondID;
-
-  if (!req.user) {
-    res.status(200).json({ friendPairsInfo: [] });
-  }
-
-  checkIfFriend(
-    reqUserID,
-    secondID,
-    (err: Error, friendPairs: FriendPairType[]) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
-
-      res.status(200).json({ friendPairsInfo: friendPairs });
-    }
-  );
-});
-
-app.post("/check-if-friend-with-user-id", (req: Request, res: Response) => {
-  const reqUserGoogleID = String(req.user?.id);
-  const secondID = req.body.secondID;
-
-  if (!req.user) {
-    res.status(200).json({ friendPairsInfo: [] });
-  }
-
-  checkIfFriendWithUserID(
-    reqUserGoogleID,
-    secondID,
-    (err: Error, friendPairs: FriendPairType[]) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
-
-      res.status(200).json({ friendPairsInfo: friendPairs });
-    }
-  );
-});
-
-// send a friend request based on google ID's
-app.post("/send-friend-request", async (req: Request, res: Response) => {
-  // check if user is logged in/google id exists
-  if (!req.user?.id) {
-    return res.status(403).json({ message: "User's Google ID not found" });
-  }
-
-  const requestorGoogleID = String(req.user?.id); // google id of person sending the request
-  const requesteeGoogleID = req.body.requesteeGoogleID; // google id of person receiving the request
-
-  sendFriendRequest(
-    requestorGoogleID,
-    requesteeGoogleID,
-    (err: Error, insertID: number) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
-
-      res.status(200).json({ insertID: insertID });
-    }
-  );
-});
-
-// accept a friend request -> insert into friends
-app.post("/accept-request", async (req: Request, res: Response) => {
-  // check if user is logged in/google id exists
-  if (!req.user?.id) {
-    return res.status(403).json({ message: "User's Google ID not found" });
-  }
-
-  const userGoogleID = String(req.user?.id);
-  const friendGoogleID = req.body.google_id;
-
-  acceptRequest(
-    userGoogleID,
-    friendGoogleID,
-    (err: Error, insertID: number) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
-
-      res.status(200).json({ insertID: insertID });
-    }
-  );
-});
-
-// deny a friend request -> delete from friend_requests
-app.post("/deny-request", async (req: Request, res: Response) => {
-  // check if user is logged in/google id exists
-  if (!req.user?.id) {
-    return res.status(403).json({ message: "User's Google ID not found" });
-  }
-
-  const userGoogleID = String(req.user?.id);
-  const friendGoogleID = req.body.google_id;
-
-  denyRequest(
-    userGoogleID,
-    friendGoogleID,
-    (err: Error, deletionID: number) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
-
-      res.status(200).json({ deletionID: deletionID });
-    }
-  );
-});
-
-// get outgoing friend requests for a user from user's google id
-app.get(
-  "/get-outgoing-friend-requests",
-  async (req: Request, res: Response) => {
-    // check if user is logged in/google id exists
-    if (!req.user?.id) {
-      return res.status(403).json({ message: "User's Google ID not found" });
-    }
-
-    const userGoogleID = String(req.user?.id);
-
-    getOutgoingFriendRequests(
-      userGoogleID,
-      (err: Error, outgoingRequestUsers: FriendRequestUserType[]) => {
-        if (err) {
-          return res.status(500).json({ message: err.message });
-        }
-
-        res.status(200).json({ outgoingRequestUsers: outgoingRequestUsers });
-      }
-    );
-  }
-);
-
-// get incoming friend requests for a user from user's google id
-app.get(
-  "/get-incoming-friend-requests",
-  async (req: Request, res: Response) => {
-    // check if user is logged in/google id exists
-    if (!req.user?.id) {
-      return res.status(403).json({ message: "User's Google ID not found" });
-    }
-    const userGoogleID = String(req.user?.id);
-
-    getIncomingFriendRequests(
-      userGoogleID,
-      (err: Error, incomingRequestUsers: FriendRequestUserType[]) => {
-        if (err) {
-          return res.status(500).json({ message: err.message });
-        }
-
-        res.status(200).json({ incomingRequestUsers: incomingRequestUsers });
-      }
-    );
-  }
-);
-
-app.get("/googleuser", (req: Request, res: Response) => {
-  res.status(200).json({ google_user: req.user?.profile });
-});
-
-// gets user id from google id
-app.get("/get-user-id", (req: Request, res: Response) => {
-  if (!req.user) {
-    return res.status(200).json({ userID: [] });
-  }
-  const userGoogleID = req.user?.profile.id;
-
-  getUserID(userGoogleID, (err: Error, userID: number) => {
-    if (err) {
-      return res.status(500).json({ message: err.message });
-    }
-
-    res.status(200).json({ userID: userID });
-  });
-});
-
-app.post("/check-if-username-exists", async (req: Request, res: Response) => {
-  const username = String(req.body.username);
-
-  checkUsername(username, (err: Error, usernameExists: string) => {
-    if (err) {
-      return res.status(500).json({ message: err.message });
-    }
-
-    res.status(200).json({ username: usernameExists });
-  });
-});
-
-app.post("/create-user", async (req: Request, res: Response) => {
-  if (!req.user?.profile.id) {
-    return res.status(500).json({ message: "No Google ID" });
-  }
-
-  const newUser = {
-    username: req.body.username,
-    first_name: req.body.firstName,
-    last_name: req.body.lastName,
-    created_at: new Date(),
-    google_id: String(req.user?.profile.id),
-    wants_to: req.body.wantsTo,
-    google_photo_link: String(req.user?.profile.photos[0].value),
-  };
-
-  createUser(newUser, (err: Error, insertID: number) => {
-    if (err) {
-      return res.status(500).json({ message: err.message });
-    }
-
-    res.status(200).json({ insertID: insertID });
-  });
-});
-
 app.get("/logout", (req: Request, res: Response, err: any) => {
   req.session.destroy((err) => {
     if (err) throw err;
     res.status(200).json({ message: "Successfuly cleared session" });
   });
 });
-
-// auth testing ----
 
 app.listen(process.env.PORT, () => {
   console.log(`Node server started running on Port ${process.env.PORT}`);
