+++ server/controllers/UserManagement.ts
@@ -1,6 +1,10 @@
 import { db } from "../db";
 import { OkPacket, RowDataPacket } from "mysql2";
-import { FriendRequestUserType, UserType } from "../types/content";
+import {
+  FriendPairType,
+  FriendRequestUserType,
+  UserType,
+} from "../types/content";
 
 export const createUser = (newUser: UserType, callback: Function) => {
   const queryString =
@@ -167,5 +171,122 @@ export const getIncomingFriendRequests = (
       incomingRequestUsers.push(incomingRequestUser);
     });
     callback(null, incomingRequestUsers);
+  });
+};
+
+// gets user id from the google id
+export const getUserID = (userGoogleID: string, callback: Function) => {
+  const queryString = "SELECT id FROM bliss_db.users WHERE google_id=?";
+
+  db.query(queryString, userGoogleID, (err, result) => {
+    if (err) {
+      console.log(err);
+    }
+
+    const userID = <RowDataPacket>result;
+    callback(null, userID);
+  });
+};
+
+// check if ID in url user is navigating to, is a friend of logged in user
+export const checkIfFriend = (
+  reqUserID: string,
+  urlID: string,
+  callback: Function
+) => {
+  // friends table is [id, user_id, friend_id] so must check if (user_id, friend_id) OR (friend_id, user_id) exists
+  const queryStringOne = "SELECT * FROM bliss_db.friends WHERE ";
+  // check if (user_id, friend_id) exists
+  const queryStringTwo =
+    "(user_id=(SELECT id FROM bliss_db.users WHERE google_id=?) AND friend_id=(SELECT id FROM bliss_db.users WHERE google_id=?))";
+  // check if "OR" (friend_id, user_id) exists
+  const queryStringThree =
+    " OR (user_id=(SELECT id FROM bliss_db.users WHERE google_id=?) AND friend_id=(SELECT id FROM bliss_db.users WHERE google_id=?))";
+
+  // combine all query strings
+  const mainQueryString: string =
+    queryStringOne + queryStringTwo + queryStringThree;
+
+  db.query(
+    mainQueryString,
+    [reqUserID, urlID, urlID, reqUserID],
+    (err, result) => {
+      if (err) {
+        callback(err);
+      }
+
+      // there should only be one, so add a check for this later
+      const rows = <RowDataPacket[]>result;
+      const friendPairs: FriendPairType[] = [];
+
+      rows.forEach((row) => {
+        const friendPair: FriendPairType = {
+          friend_one_id: row.user_id,
+          friend_two_id: row.friend_id,
+        };
+        friendPairs.push(friendPair);
+      });
+      callback(null, friendPairs);
+    }
+  );
+};
+
+export const checkIfFriendWithUserID = (
+  reqUserGoogleID: string,
+  secondID: string,
+  callback: Function
+) => {
+  // friends table is [id, user_id, friend_id] so must check if (user_id, friend_id) OR (friend_id, user_id) exists
+  const queryStringOne = "SELECT * FROM bliss_db.friends WHERE ";
+  // check if (user_id, friend_id) exists
+  const queryStringTwo =
+    "(user_id=(SELECT id FROM bliss_db.users WHERE google_id=?) AND friend_id=?)";
+  // check if "OR" (friend_id, user_id) exists
+  const queryStringThree =
+    " OR (user_id=? AND friend_id=(SELECT id FROM bliss_db.users WHERE google_id=?))";
+
+  // combine all query strings
+  const mainQueryString: string =
+    queryStringOne + queryStringTwo + queryStringThree;
+
+  db.query(
+    mainQueryString,
+    [reqUserGoogleID, secondID, secondID, reqUserGoogleID],
+    (err, result) => {
+      if (err) {
+        callback(err);
+      }
+
+      // there should only be one, so add a check for this later
+      const rows = <RowDataPacket[]>result;
+      const friendPairs: FriendPairType[] = [];
+
+      rows.forEach((row) => {
+        const friendPair: FriendPairType = {
+          friend_one_id: row.user_id,
+          friend_two_id: row.friend_id,
+        };
+        friendPairs.push(friendPair);
+      });
+      callback(null, friendPairs);
+    }
+  );
+};
+
+// update google photo based on google id since you can't select "FROM" the table you're updating
+export const updateGooglePhoto = (
+  userGoogleID: string,
+  googlePhotoLink: string,
+  callback: Function
+) => {
+  const queryString = "UPDATE users SET google_photo_link=? WHERE google_id=?";
+
+  db.query(queryString, [googlePhotoLink, userGoogleID], (err, result) => {
+    if (err) {
+      callback(err);
+    }
+
+    const insertID = (<OkPacket>result).insertId;
+    callback(null, insertID);
   });
 };
