+++ server/routes/ManageContent.ts
@@ -1,21 +1,21 @@
 import express, { Request, Response } from "express";
+import Query from "mysql2/typings/mysql/lib/protocol/sequences/Query";
 import {
-  createBucketList,
+  addBucketList,
+  updateBucketList,
+  deleteBucketList,
   addActivity,
   updateActivity,
   deleteActivity,
-  updateBucketList,
-  deleteBucketList,
   addSharedListUsers,
   removeSharedListUsers,
-  updateGooglePhoto,
 } from "../controllers/ManageContent";
 import { BucketList, BucketListContent } from "../types/content";
 
 const contentRouter = express.Router();
 
-// create new bucket list
-contentRouter.post("/create", async (req: Request, res: Response) => {
+// add new bucket list
+contentRouter.post("/bucket-list", async (req: Request, res: Response) => {
   const newBucketList: BucketList = {
     google_id: req.user?.profile.id,
     privacy_type: req.body.privacy_type,
@@ -24,7 +24,7 @@ contentRouter.post("/create", async (req: Request, res: Response) => {
     permissions: req.body.permissions,
   };
 
-  createBucketList(newBucketList, (err: Error, creationId: number) => {
+  addBucketList(newBucketList, (err: Error, creationId: number) => {
     if (err) {
       return res.status(500).json({ message: err.message });
     }
@@ -33,8 +33,40 @@ contentRouter.post("/create", async (req: Request, res: Response) => {
   });
 });
 
+// delete a bucket list
+contentRouter.delete("/bucket-list", async (req: Request, res: Response) => {
+  const trackerID = Number(req.params.id);
+
+  deleteBucketList(trackerID, (err: Error, insertId: number) => {
+    if (err) {
+      return res.status(500).json({ message: err.message });
+    }
+
+    res.status(200).json({ insertId: insertId });
+  });
+});
+
+// update a bucket list's info
+contentRouter.patch("/bucket-list", async (req: Request, res: Response) => {
+  const updatedBucketList: BucketList = {
+    id: req.body.id,
+    privacy_type: req.body.privacy_type,
+    title: req.body.title,
+    description: req.body.description,
+    permissions: req.body.permissions,
+  };
+
+  updateBucketList(updatedBucketList, (err: Error, updateID: number) => {
+    if (err) {
+      return res.status(500).json({ message: err.message });
+    }
+
+    res.status(200).json({ updateID: updateID });
+  });
+});
+
 // add an activity to a particular bucket list based on id of that list
-contentRouter.post("/add", async (req: Request, res: Response) => {
+contentRouter.post("/activity", async (req: Request, res: Response) => {
   const newActivity: BucketListContent = req.body;
 
   addActivity(newActivity, (err: Error, addId: number) => {
@@ -46,40 +78,8 @@ contentRouter.post("/add", async (req: Request, res: Response) => {
   });
 });
 
-// delete an activity from a bucket list based on id of that list
-contentRouter.post("/delete-activity", async (req: Request, res: Response) => {
-  const activityIDs = {
-    trackerID: req.body.tracker_id,
-    contentID: req.body.content_id,
-  };
-
-  deleteActivity(activityIDs, (err: Error, insertID: number) => {
-    if (err) {
-      return res.status(500).json({ message: err.message });
-    }
-
-    res.status(200).json({ insertID: insertID });
-  });
-});
-
-// delete a bucket list
-contentRouter.post(
-  "/delete-bucket-list",
-  async (req: Request, res: Response) => {
-    const trackerID = Number(req.body.id);
-
-    deleteBucketList(trackerID, (err: Error, insertId: number) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
-
-      res.status(200).json({ insertId: insertId });
-    });
-  }
-);
-
 // update an activity in a bucket list
-contentRouter.put("/update-activity", async (req: Request, res: Response) => {
+contentRouter.patch("/activity", async (req: Request, res: Response) => {
   const updatedActivity: BucketListContent = req.body;
 
   updateActivity(updatedActivity, (err: Error, updatedId: number) => {
@@ -91,43 +91,31 @@ contentRouter.put("/update-activity", async (req: Request, res: Response) => {
   });
 });
 
-// update a bucket list's info
-contentRouter.put(
-  "/update-bucket-list",
-  async (req: Request, res: Response) => {
-    const updatedBucketList: BucketList = {
-      id: req.body.id,
-      privacy_type: req.body.privacy_type,
-      title: req.body.title,
-      description: req.body.description,
-      permissions: req.body.permissions,
-    };
+// delete an activity from a bucket list based on id of that list
+contentRouter.delete("/activity", async (req: Request, res: Response) => {
+  const activityIDs = {
+    trackerID: Number(req.query.tracker_id),
+    contentID: Number(req.query.content_id),
+  };
 
-    updateBucketList(updatedBucketList, (err: Error, updateID: number) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
+  deleteActivity(activityIDs, (err: Error, insertID: number) => {
+    if (err) {
+      return res.status(500).json({ message: err.message });
+    }
 
-      res.status(200).json({ updateID: updateID });
-    });
-  }
-);
+    res.status(200).json({ insertID: insertID });
+  });
+});
 
 // add various users to shared_list_users from supplied user ID's
 contentRouter.post(
-  "/add-shared-list-users",
+  "/shared-list-users",
   async (req: Request, res: Response) => {
-    const ownerID = Number(req.body.ownerID);
     const bucketListID = Number(req.body.bucketListID);
     const selectedUserIDs: number[] = req.body.selectedUserIDs;
 
     // if anything is empty, return
-    if (
-      !ownerID ||
-      !bucketListID ||
-      !selectedUserIDs ||
-      selectedUserIDs.length === 0
-    ) {
+    if (!bucketListID || !selectedUserIDs || selectedUserIDs.length === 0) {
       return res
         .status(400)
         .json({ message: "Not all necessary inputs were sent." });
@@ -137,7 +125,7 @@ contentRouter.post(
     let convertedArray: number[][] = [];
 
     for (let i = 0; i < selectedUserIDs.length; i++) {
-      convertedArray.push([bucketListID, selectedUserIDs[i], ownerID]);
+      convertedArray.push([bucketListID, selectedUserIDs[i]]);
     }
 
     addSharedListUsers(convertedArray, (err: Error, insertID: number) => {
@@ -152,10 +140,10 @@ contentRouter.post(
 
 // remove various users to shared_list_users from supplied user ID's
 contentRouter.post(
-  "/remove-shared-list-users",
+  "/delete-shared-list-users",
   async (req: Request, res: Response) => {
     const bucketListID = Number(req.body.bucketListID);
-    const removedUserIDs: number[] = req.body.removedUserIDs;
+    const removedUserIDs = req.body.removedUserIDs;
 
     // if anything is empty, return
     if (!bucketListID || !removedUserIDs || removedUserIDs.length === 0) {
@@ -181,29 +169,7 @@ contentRouter.post(
   }
 );
 
-// update profile photo to reflect new google photo
-contentRouter.put(
-  "/update-google-photo",
-  async (req: Request, res: Response) => {
-    // check if user is logged in/google id exists
-    if (!req.user?.id) {
-      return res.status(403).json({ message: "User's Google ID not found" });
-    }
-    const userGoogleID = String(req.user?.id);
-    const googlePhotoLink = String(req.body.googlePhotoLink);
-
-    updateGooglePhoto(
-      userGoogleID,
-      googlePhotoLink,
-      (err: Error, insertID: number) => {
-        if (err) {
-          return res.status(500).json({ message: err.message });
-        }
-
-        res.status(200).json({ insertID: insertID });
-      }
-    );
-  }
-);
+// updates wants_to portion of user
+// contentRouter.patch("/")
 
 export { contentRouter };
