+++ server/routes/ViewContent.ts
@@ -44,7 +44,7 @@ viewContentRouter.get(
 
 // get only public_random lists for viewing profiles where user is not friends with them
 viewContentRouter.get(
-  "/get-public-lists/:id",
+  "/public-lists/:id",
   async (req: Request, res: Response) => {
     const userGoogleID = req.params.id;
 
@@ -60,27 +60,24 @@ viewContentRouter.get(
 
 // get lists for friend profile, get all public_friends/public_random and relevant shared lists
 // note, this gets called after "/get-shared-lists/:id" gets called down below to get shared list ID's
-viewContentRouter.post(
-  "/get-friend-lists",
-  async (req: Request, res: Response) => {
-    const sharedListIDs = {
-      array: req.body?.sharedListArray,
-      friendGoogleID: req.body.friendGoogleID,
-    };
+viewContentRouter.post("/friend-lists", async (req: Request, res: Response) => {
+  const sharedListIDs = {
+    array: req.body?.sharedListArray,
+    friendGoogleID: req.body.friendGoogleID,
+  };
 
-    getFriendsLists(
-      sharedListIDs.array,
-      sharedListIDs.friendGoogleID,
-      (err: Error, lists: BucketList[]) => {
-        if (err) {
-          return res.status(500).json({ message: err.message });
-        }
-
-        res.status(200).json({ data: lists });
+  getFriendsLists(
+    sharedListIDs.array,
+    sharedListIDs.friendGoogleID,
+    (err: Error, lists: BucketList[]) => {
+      if (err) {
+        return res.status(500).json({ message: err.message });
       }
-    );
-  }
-);
+
+      res.status(200).json({ data: lists });
+    }
+  );
+});
 
 // get title and description of a bucket list
 viewContentRouter.get(
@@ -139,7 +136,7 @@ viewContentRouter.get(
 
 // check if user is in shared list for a particular bucket list
 viewContentRouter.get(
-  "/check-if-user-in-shared-list/:id",
+  "/shared-list-user/:id",
   async (req: Request, res: Response) => {
     const userID = String(req.user?.id);
     const bucketListID = Number(req.params.id);
@@ -159,29 +156,26 @@ viewContentRouter.get(
 
 // check if user, who is visiting friend profile, is in any shared_list_users rows with the friend and return those bucket_list_id's
 // (uses Google ID)
-viewContentRouter.get(
-  "/get-shared-lists/:id",
-  (req: Request, res: Response) => {
-    const userGoogleID = req.user?.profile.id;
-    const friendGoogleID = req.params.id;
+viewContentRouter.get("/shared-lists/:id", (req: Request, res: Response) => {
+  const userGoogleID = req.user?.profile.id;
+  const friendGoogleID = req.params.id;
 
-    getSharedLists(
-      userGoogleID,
-      friendGoogleID,
-      (err: Error, bucketListIDs: number[]) => {
-        if (err) {
-          return res.status(500).json({ message: err.message });
-        }
-
-        res.status(200).json({ bucketListIDs: bucketListIDs });
+  getSharedLists(
+    userGoogleID,
+    friendGoogleID,
+    (err: Error, bucketListIDs: number[]) => {
+      if (err) {
+        return res.status(500).json({ message: err.message });
       }
-    );
-  }
-);
+
+      res.status(200).json({ bucketListIDs: bucketListIDs });
+    }
+  );
+});
 
 // get all users in a shared list based on provided bucket list tracker id (id)
 viewContentRouter.get(
-  "/get-shared-list-users/:id",
+  "/all-shared-list-users/:id",
   async (req: Request, res: Response) => {
     const trackerID = Number(req.params.id);
 
@@ -197,7 +191,7 @@ viewContentRouter.get(
 
 // get all contributor's for all of user's owned, shared bucket lists
 viewContentRouter.get(
-  "/get-all-contributors",
+  "/all-contributors",
   async (req: Request, res: Response) => {
     // check if user is logged in/google id exists
     if (!req.user?.id) {
@@ -219,24 +213,21 @@ viewContentRouter.get(
 );
 
 // get user info from database from google ID
-viewContentRouter.get(
-  "/get-user-info/:id",
-  async (req: Request, res: Response) => {
-    const googleID = req.params.id;
+viewContentRouter.get("/user-info/:id", async (req: Request, res: Response) => {
+  const googleID = req.params.id;
 
-    getUserInfo(googleID, (err: Error, userInfo: any[]) => {
-      if (err) {
-        return res.status(500).json({ message: err.message });
-      }
+  getUserInfo(googleID, (err: Error, userInfo: any[]) => {
+    if (err) {
+      return res.status(500).json({ message: err.message });
+    }
 
-      res.status(200).json({ userInfo: userInfo });
-    });
-  }
-);
+    res.status(200).json({ userInfo: userInfo });
+  });
+});
 
 // get list of friends from google id
 viewContentRouter.get(
-  "/get-list-of-friends",
+  "/list-of-friends",
   async (req: Request, res: Response) => {
     // check if user is logged in/google id exists
     if (!req.user?.id) {
@@ -257,7 +248,7 @@ viewContentRouter.get(
 
 // get full list of users for search list, excluding current user
 viewContentRouter.get(
-  "/get-full-user-list/:id",
+  "/full-user-list/:id",
   async (req: Request, res: Response) => {
     const userGoogleID = String(req.params.id);
     getUserList(userGoogleID, (err: Error, userList: FullUserListType[]) => {
