+++ server/types/content.ts
@@ -34,7 +34,6 @@ export interface SharedListUserType {
   id?: number;
   bucket_list_id: number;
   contributor_id: number;
-  owner_id?: number;
 }
 
 export interface UserType {
